Fix deprecated "implicit block" expectation syntax
RSpec wants us to use an explicit block rather than passing in a proc.
In the Counter spec, this was a straightforward syntax change.

In the Beefcake Protobuffs backend spec, I just removed the expectation
of no errors raised. That's generally what all of our tests expect
anyway. If an error is raised, it'll show up as a test failure.